book search option to navbar and fix unnecessary book info retrieval bug
