+++ book_search.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="main.css">
+    <script src="https://code.jquery.com/jquery-3.6.1.js"></script>
+    <script type="module" src="js/navbar.js"></script>
+    <script type="module" src="js/search_book.js"></script>
+
+    <title>Alexandria - Personal Listings</title>
+</head>
+<body>
+<!-- navbar for website navigation -->
+<div class="navbar">
+    <div class="navbar">
+        <div class="container">
+            <ul class="left-nav">
+                <li><a class="logo" href="index.html">Alexandria</a></li>
+                <span id="other-left-navbar-items-toggle"></span>
+                <li><a href="create_account.html">Create Account</a></li>
+                <li id="login-or-logout-toggle"></li>
+            </ul>
+
+            <ul class="right-nav">
+                <li id="profile-username-toggle"></li>
+                <li class="token-display" id="token-display-toggle"></li>
+            </ul>
+        </div>
+    </div>
+</div>
+
+<!-- header -->
+<header>
+    <h1>Search for Book</h1>
+</header>
+
+ <!-- search information section -->
+ <section class="search_info">
+    <div class="container">
+        <form action="#">
+            <label id="book-search" for="book-search">Search for Book</label><br>
+            <label class="desc" id="university-desc">Enter a book name or isbn</label>
+            <input type="text" id="book-search" value=""><br>
+
+            <input type="submit" class="submit-button" id="submit-button" value="Search">
+            <p id="submit-message"></p>
+        </form>
+    </div>
+</section>
+
+<!-- search results section -->
+<section class="book-info">
+    <div class="container">
+        <form action="#">
+            <div class="book-collection">
+                <table class="collection-list">
+                    <tbody>
+
+                    </tbody>
+                </table>
+            </div>
+        </form>
+    </div>
+</section>
+
+</body>
+</html>
