+++ js/vars_and_helpers.js
@@ -1,10 +1,13 @@
 const localSpringBootServerUrl = "http://localhost:8080";
 const highlightInputFieldColor = "rgb(255, 0, 0)";
 const minimumBookDescriptionLength = 10, maximumBookDescriptionLength = 250;
-const indexPageFilename = "index.html", loginPageFilename = "login.html", listBookPageFilename = "list_book.html", personalListingsPageFilename = "personal_listings.html";
+const indexPageFilename = "index.html", loginPageFilename = "login.html", 
+    listBookPageFilename = "list_book.html", personalListingsPageFilename = "personal_listings.html",
+    bookSearchPageFilename = "book_search.html";
 
-const otherLeftNavbarItems = '<li><a href="list_book.html">List Book</a></li> \
-<li><a href="personal_listings.html">My Listings</a></li>';
+const otherLeftNavbarItems = `<li><a href=${listBookPageFilename}>List Book</a></li>` + 
+`<li><a href=${personalListingsPageFilename}>My Listings</a></li>` + 
+`<li><a href=${bookSearchPageFilename}>Search for Book</a></li>`;
 
 const checkStringNotEmpty = x => x != null && x.length > 0;
 
@@ -96,10 +99,39 @@ const highlightText = (ele, newColor) => ele.css("color", newColor);
 
 const resetStyle = element => element.attr("style", "");
 
+const displayBooks = () => {
+    var bookList = JSON.parse(window.sessionStorage.getItem("bookList"));
+    if (!bookList || bookList.length === 0) {
+        //If No Books
+    }
+    else {
+        for (var book of bookList) {
+            $('.collection-list').find('tbody').append(
+                '<tr class="collection-row">'+
+                    '<td>'+
+                        '<button class="collection-row-content">'+
+                            ' <div class="list-div">'+
+                                '<img src="images/book-add.png" class="collection-book-image">'+
+                                //'<img src="images/minus-button.png" class="remove-element-button">'+
+                                `<div>${book.name}</div>`+
+                                `<div class="collection-book-attr">ISBN: ${book.isbn}</div>`+
+                                `<div class="collection-book-attr">Condition: ${book.condition}</div>`+
+                                `<div class="collection-book-attr">${(book.forExchange) ? "For Exchange" : "For Give Away"}</div>`+
+                                `<div class="collection-book-attr">Description: ${book.description}</div>`+
+                                '</div>'+
+                            '</button>'+
+                        '</td>'+
+                   '</tr>'
+            );
+        }
+    }
+};
+
 export { localSpringBootServerUrl, minimumBookDescriptionLength, maximumBookDescriptionLength, 
     indexPageFilename, loginPageFilename, listBookPageFilename, personalListingsPageFilename,
     otherLeftNavbarItems,
     checkStringNotEmpty, checkEmail, checkPassword, checkISBN, 
     checkBookDescriptionNotTooShort, checkBookDescriptionNotTooLong, 
-    highlightInputField, highlightText, resetStyle };
+    highlightInputField, highlightText, resetStyle,
+    displayBooks };
